+++ CHANGELOG.md
@@ -9,6 +9,7 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 - upcloud_file_storage_share: fix resource import by parsing the file storage service uuid and share name from id
 - upcloud_file_storage_share_acl: fix resource import by parsing the file storage service uuid, share name and acl name from id
+- upcloud_managed_database_*: do not plan replacement for unconfigured create-only properties, such as MySQL `properties.lower_case_table_names`
 
 ## [5.36.1] - 2026-03-26